+++ R/sentencepiece.R
@@ -131,7 +131,20 @@ sentencepiece_load_model <- function(file = "sentencepiece.model"){
 #' @param x a character vector of text (in UTF-8 Encoding)
 #' @param type a character string, either 'subwords' or 'ids' to get the subwords or the corresponding ids of these subwords as defined in the vocabulary of the model. 
 #' Defaults to 'subwords'.
-#' @return a list with tokenised text, one for each element of \code{x}
+#' @param nbest integer indicating the number of segmentations to extract. See the details. The argument is not used if you do not provide a value for it.
+#' @param alpha smoothing parameter to perform subword regularisation. Typical values are 0.1, 0.2 or 0.5. See the details. The argument is not used if you do not provide a value for it or do not provide a value for \code{nbest}.
+#' @return a list with tokenised text, one for each element of \code{x} 
+#' unless you provide \code{nbest} without providing \code{alpha} in which case the result is a list of list of \code{nbest} tokenised texts 
+#' @details 
+#' If you specify \code{alpha} to perform subword regularisation, keep in mind the following. \cr
+#' When alpha is 0.0, one segmentation is uniformly sampled from the \code{nbest} or lattice. 
+#' The best Viterbi segmentation is more likely sampled when setting larger \code{alpha} values like 0.1. \cr
+#' \itemize{
+#' \item If you provide a positive value for \code{nbest}, approximately samples one segmentation from \code{nbest} candidates.
+#' \item If you provide a negative value for \code{nbest}, samples one segmentation from the hypotheses (Lattice) according to the generation probabilities using forward-filtering and backward-sampling algorithm.
+#' }
+#' \code{nbest} and \code{alpha} correspond respectively to the parameter \code{l} and in \code{alpha} 
+#' in the paper \url{https://arxiv.org/abs/1804.10959} where (\code{nbest} < 0 means l = infinity).\cr
 #' @export
 #' @examples
 #' model <- system.file(package = "sentencepiece", "models", "nl-fr-dekamer.model")
@@ -141,16 +154,58 @@ sentencepiece_load_model <- function(file = "sentencepiece.model"){
 #'          "On est d'accord sur le prix de la biere?")
 #' sentencepiece_encode(model, x = txt, type = "subwords")
 #' sentencepiece_encode(model, x = txt, type = "ids")
-sentencepiece_encode <- function(model, x, type = c("subwords", "ids")){
+#' 
+#' ## Examples using subword regularisation
+#' model <- system.file(package = "sentencepiece", "models", "nl-fr-dekamer-unigram.model")
+#' model <- sentencepiece_load_model(file = model)
+#' 
+#' txt <- c("Goed zo",
+#'          "On est d'accord")
+#' sentencepiece_encode(model, x = txt, type = "subwords", nbest = 4)
+#' sentencepiece_encode(model, x = txt, type = "ids", nbest = 4)
+#' sentencepiece_encode(model, x = txt, type = "subwords", nbest = 2)
+#' sentencepiece_encode(model, x = txt, type = "ids", nbest = 2)
+#' sentencepiece_encode(model, x = txt, type = "subwords", nbest = 1)
+#' sentencepiece_encode(model, x = txt, type = "ids", nbest = 1)
+#' sentencepiece_encode(model, x = txt, type = "subwords", nbest = 4, alpha = 0.1)
+#' sentencepiece_encode(model, x = txt, type = "ids", nbest = 4, alpha = 0.1)
+#' sentencepiece_encode(model, x = txt, type = "subwords", nbest = -1, alpha = 0.1)
+#' sentencepiece_encode(model, x = txt, type = "ids", nbest = -1, alpha = 0.1)
+#' sentencepiece_encode(model, x = txt, type = "subwords", nbest = -1, alpha = 0)
+#' sentencepiece_encode(model, x = txt, type = "ids", nbest = -1, alpha = 0)
+sentencepiece_encode <- function(model, x, type = c("subwords", "ids"), nbest = -1L, alpha = 0.1){
   type <- match.arg(type)
   if(type == "ids"){
-    x <- spc_encode_as_ids(model$model, x)  
+    if(!missing(nbest) && !missing(alpha)){
+      x <- spc_encode_as_ids_sample(model$model, x, as.integer(nbest), as.numeric(alpha))  
+    }else if(!missing(nbest)){
+      x <- spc_encode_as_ids_nbest(model$model, x, as.integer(nbest))  
+    }else{
+      x <- spc_encode_as_ids(model$model, x)  
+    }
   }else if(type == "subwords"){
-    x <- spc_encode_as_subwords(model$model, x)  
-    x <- lapply(x, FUN=function(x){
-      Encoding(x) <- "UTF-8"
-      x
-    })
+    if(!missing(nbest) && !missing(alpha)){
+      x <- spc_encode_as_subwords_sample(model$model, x, as.integer(nbest), as.numeric(alpha))  
+      x <- lapply(x, FUN=function(x){
+        Encoding(x) <- "UTF-8"
+        x
+      })
+    }else if(!missing(nbest)){
+      x <- spc_encode_as_subwords_nbest(model$model, x, as.integer(nbest))  
+      x <- lapply(x, FUN=function(x){
+        x <- lapply(x, FUN=function(x){
+          Encoding(x) <- "UTF-8"
+          x
+        })
+        x
+      })
+    }else{
+      x <- spc_encode_as_subwords(model$model, x)  
+      x <- lapply(x, FUN=function(x){
+        Encoding(x) <- "UTF-8"
+        x
+      })
+    }
   }
   x
 }
@@ -163,7 +218,7 @@ sentencepiece_encode <- function(model, x, type = c("subwords", "ids")){
 #' @param model an object of class \code{sentencepiece} as returned by \code{\link{sentencepiece_load_model}} or \code{\link{sentencepiece}}
 #' @param x an integer vector of Sentencepiece id's or a list of these
 #' @export
-#' @return a character vector of detokenised text
+#' @return a character vector of detokenised text or if you encoded with \code{nbest}, a list of these
 #' @examples 
 #' model <- system.file(package = "sentencepiece", "models", "nl-fr-dekamer.model")
 #' model <- sentencepiece_load_model(file = model)
@@ -175,23 +230,38 @@ sentencepiece_encode <- function(model, x, type = c("subwords", "ids")){
 #' sentencepiece_decode(model, x)
 #' x <- sentencepiece_encode(model, x = txt, type = "ids")
 #' sentencepiece_decode(model, x)
+#' 
+#' model <- system.file(package = "sentencepiece", "models", 
+#'                      "nl-fr-dekamer-unigram.model")
+#' model <- sentencepiece_load_model(file = model)
+#' x <- sentencepiece_encode(model, x = txt, type = "subwords", nbest = 3)
+#' sentencepiece_decode(model, x)
+#' x <- sentencepiece_encode(model, x = txt, type = "subwords", 
+#'                           nbest = 3, alpha = 0.1)
+#' sentencepiece_decode(model, x)
 sentencepiece_decode <- function(model, x){
   if(is.integer(x) || (is.list(x) && is.integer(x[[1]]))){
     f <- spc_decode_ids
   }else{
     f <- spc_decode_subwords
   }
-  if(!is.list(x)){
-    x <- f(model$model, x) 
-    Encoding(x) <- "UTF-8"
-  }else{
+  if(is.list(x) && is.list(x[[1]])){
     x <- lapply(x, FUN=function(x){
-      x <- f(model$model, x)
-      Encoding(x) <- "UTF-8"
-      x
+      sentencepiece_decode(model, x)  
     })
+  }else{
+    if(!is.list(x)){
+      x <- f(model$model, x) 
+      Encoding(x) <- "UTF-8"
+    }else{
+      x <- lapply(x, FUN=function(x){
+        x <- f(model$model, x)
+        Encoding(x) <- "UTF-8"
+        x
+      })
+    }
+    x <- unlist(x)
   }
-  x <- unlist(x)
   x
 }